# Conflicts:
#	Assets/!BratsStuff/BRATANDRONIKPROJECTS/WhoHeroes.meta
#	Assets/!BratsStuff/BRATANDRONIKPROJECTS/WhoHeroes/!Prefabs.meta
#	Assets/!BratsStuff/BRATANDRONIKPROJECTS/WhoHeroes/!Prefabs/Scene.meta
#	Assets/!BratsStuff/BRATANDRONIKPROJECTS/WhoHeroes/!Prefabs/Scene/WhoHeroes_MainLocals.prefab
#	Assets/!BratsStuff/BRATANDRONIKPROJECTS/WhoHeroes/!Prefabs/Scene/WhoHeroes_Transforms.prefab
#	Assets/Layer Lab/GUI Pro-FantasyRPG.meta
#	Assets/Layer Lab/GUI Pro-FantasyRPG/ResourcesData.meta
#	Assets/Layer Lab/GUI Pro-FantasyRPG/ResourcesData/Fonts.meta
#	Assets/Layer Lab/GUI Pro-FantasyRPG/ResourcesData/Fonts/Alata-Regular SDF.asset
#	Assets/Layer Lab/GUI Pro-FantasyRPG/ResourcesData/Fonts/Alata-Regular SDF.asset.meta
#	Assets/Layer Lab/GUI Pro-FantasyRPG/ResourcesData/Sprites.meta
#	Assets/Layer Lab/GUI Pro-FantasyRPG/ResourcesData/Sprites/Component.meta
#	Assets/Layer Lab/GUI Pro-FantasyRPG/ResourcesData/Sprites/Component/Frame.meta
#	Assets/Layer Lab/GUI Pro-FantasyRPG/ResourcesData/Sprites/Component/Frame/BasicFrame_Diamond_80_Glow.png.meta
#	Assets/Layer Lab/GUI Pro-FantasyRPG/ResourcesData/Sprites/Component/Frame/ItemFrame_02_Glow.png.meta
#	Assets/Layer Lab/GUI Pro-FantasyRPG/ResourcesData/Sprites/Component/Frame/LineTextFrame_03_Line.png.meta
#	Assets/Layer Lab/GUI Pro-FantasyRPG/ResourcesData/Sprites/Component/Icon_PictoIcons.meta
#	Assets/Layer Lab/GUI Pro-FantasyRPG/ResourcesData/Sprites/Component/Icon_PictoIcons/512.meta
#	Assets/Layer Lab/GUI Pro-FantasyRPG/ResourcesData/Sprites/Demo.meta
#	Assets/Layer Lab/GUI-TheStone.meta
#	Assets/Layer Lab/GUI-TheStone/ResourcesData.meta
#	Assets/Layer Lab/GUI-TheStone/ResourcesData/Sprites.meta
#	Assets/Layer Lab/GUI-TheStone/ResourcesData/Sprites/Components.meta
#	Assets/StreamingAssets/WhoHeroes/Config_whoheroes.xlsx
#	Assets/StreamingAssets/WhoHeroes/Config_whoheroes.xlsx.meta
#	Assets/TextMesh Pro/Resources/Fonts & Materials/LiberationSans SDF - Fallback.asset

+++ Assets/System/Prefabs Extra/XDbuff.cs
@@ -60,7 +60,7 @@ public class XDbuff : ComponentBehavior
             var ii = transform.GetChild(i);
             var bb = mon.timedBuffs[i];
             ii.Find("icon").GetComponent<Image>().sprite = ResourceHolder.instance.skills[bb.dbObj.ID];
-            ii.Find("txt").GetComponent<TextMeshProUGUI>().text = bb.dbObj.ID + "(" + bb.GetPar("timeLeft") + ")";
+            ii.Find("txt").GetComponent<TextMeshProUGUI>().text = bb.dbObj.ID + "(" + (int)bb.GetPar("timeLeft") + ")";
         }
 
     }
